fix to retrieve the correct TestSet to Execute
Using FindTestSet method it retrieve the TestSetList of all the subfolders. So if you have more testset under different subfolders from a giving path with the same name the plugin should run the wrong testset. So I added the condition (at line 421) that the nodeID of the parent folder must be the same of the last folder of the given path.